+++ src/main/resources/io/jenkins/plugins/orka/PortMapper/config.jelly
@@ -1,0 +1,20 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:f="/lib/form" xmlns:c="/lib/credentials" xmlns:s="/io/jenkins/temp/jelly" xmlns:orka="/lib/orka">
+  <orka:blockWrapper>
+
+    <f:entry title="${%Node Port}" field="from">
+      <f:number min="1024" max="65535"/>
+    </f:entry>
+
+    <f:entry title="${%VM Port}" field="to">
+      <f:number/>
+    </f:entry>
+
+    <f:entry title="">
+      <div align="right">
+        <f:repeatableDeleteButton value="${%Delete Mapping}"/>
+      </div>
+    </f:entry>
+
+  </orka:blockWrapper>
+</j:jelly>
